round-4: full DAHDSR envelope + pitch bend + aftertouch + RPN 0
- Voice trait gains optional set_pitch_bend_cents() and set_pressure()
  defaults so SF2 / tone voices opt into pitch + pressure modulation
  without breaking the round-3 contract.

- Sf2Voice now drives a full DAHDSR envelope (delay, attack, hold,
  exponential decay-to-sustain, sustain, exponential release) sourced
  from SF2 generators 33-38 (delayVolEnv / attackVolEnv / holdVolEnv /
  decayVolEnv / sustainVolEnv / releaseVolEnv). Generators in
  timecents (`2^(tc/1200)` seconds); sustainVolEnv is centibels (10 cB
  = 1 dB). Spec-default `i32::MIN` sentinels fall back to round-4
  musical defaults (5 ms attack, 100 ms decay, full sustain, 100 ms
  release) so banks without explicit envelope generators still get an
  audible ADSR shape rather than the round-3 flat AR. Mid-attack
  releases capture the live envelope level so a quick note-off from
  attack doesn't pop to peak. ToneVoice picks up the same DAHDSR
  shape with a 70 % sustain level so the test fixture is now
  measurably more dynamic than round-3's sustain-at-peak AR.

- Pitch bend (status `En`): per-channel 14-bit scalar (centre 0x2000)
  → cents via `pitch_bend_to_cents(value, range)`. Mixer pushes the
  cents value into every still-held voice on the channel via
  `Voice::set_pitch_bend_cents`; new note-ons inherit the live bend so
  notes started while the wheel is held bend on their first sample.
  RPN 0 (CC 100/101 = 0/0, then CC 6 = semitones / CC 38 = cents)
  rewrites `pitch_bend_range_cents` and re-applies the live bend to
  held voices. RPN-null (default 0x3FFF) blocks stray CC 6 / 38 from
  clobbering the range.

- Channel pressure (status `Dn`) and poly aftertouch (status `An`):
  routed to `Voice::set_pressure` (default modulator → volume per
  SF2 §8.4.x). Channel pressure hits every voice on the channel; poly
  pressure only the matching key. Default voice gain curve is
  `1.0 + 0.5 * pressure` (rest 1.0 → full 1.5) so pressure boosts
  rather than gates.

- Default modulator coverage (the General-MIDI essentials): velocity
  → initialAttenuation (folded into voice amplitude via the existing
  v² curve plus the new gen-48 attenuation in centibels), aftertouch
  → volume (set_pressure), pitch wheel → pitch (set_pitch_bend_cents).
  Modwheel → vibrato is out of round-4 scope (no LFO yet).

- Tests grow 71 → 95 (+24): timecents/centibels conversion, ADSR
  shape (decay phase distinct from attack peak), pitch-bend wiring
  through mixer + held-voice + at-note-on paths, channel/poly pressure
  routing, RPN 0 data-entry, mid-attack release level capture, full
  DAHDSR override via SF2 generators, end-to-end SMF fixtures with a
  pitch-bend event and a channel-pressure event mid-note.

- Drive-by: oxideav-core "0.2" → "0.1" in Cargo.toml (matches every
  other crate in the workspace; the "0.2" was an upstream typo —
  workspace patches all redirect through 0.1.10) and a one-byte
  `as u32` cast cleanup in smf.rs test code so `clippy --all-targets`
  is clean.

+++ src/smf.rs
@@ -629,7 +629,7 @@ mod tests {
         let mut buf = vec![v & 0x7F];
         v >>= 7;
         while v != 0 {
-            buf.push(((v & 0x7F) | 0x80) as u32);
+            buf.push((v & 0x7F) | 0x80);
             v >>= 7;
         }
         // we built it LSB-first; reverse to get MSB-first wire order.
